Split runtime handles out of GC roots
